build: add a new script helper
Summary:
This script can be used to run all examples that have been build with
meson, this is usefull in order to check if everything works, and checks
if (due to some patches) errors are happening in the output.

This is not meant as function checking of the whole example, just
automatically checking the first-look-appearance.

For now a few examples are setted up correctly, there might be more that
need setups / custom arguments, they can be added over time.

Reviewers: stefan_schmidt, zmike, cedric, q66, devilhorns, segfaultxavi

Reviewed By: zmike

Subscribers: cedric, #reviewers, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D7979