efl_ui_focus_object: fixup properties that are used for event infos.
Summary:
those are the completly wrong properties, please be carefull with this.
There is a major difference between focus_geometry and geometry on the
entity. In a viewport for example its the viewport itself - the bars at
the side. Additionally, not every focus object is a efl.gfx.entity.

this fixes a giant amount of errors when you try to test terminology,
emixer, toolbar tests or something like this.

fixup 1d9fef8da6fb4f679113d9dc4f57f1520fd31936

Reviewers: cedric, segfaultxavi, zmike

Reviewed By: cedric

Subscribers: #reviewers, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D8300