efl: mark render,post event as beta.
Summary:
render,post is just forwarding the legacy type which is not the best as it expose
internal structure directly. Planning to move to an accessor for after this release,
but at this time it does imply way to much change for this current release.

Depends on D8249

Reviewers: zmike, stefan_schmidt, segfaultxavi, bu5hm4n

Reviewed By: zmike, bu5hm4n

Subscribers: #reviewers, #committers

Tags: #efl_api, PHID-PROJ-55rnlag4d454jfmlmuhu

Maniphest Tasks: T7729

Differential Revision: https://phab.enlightenment.org/D8250