efl.pack_linear: Clarify behavior and docs
Some APIs accept both positive and negative indices when accessing items.
This patch changes the documentation for the lower limit from `-(count - 1)` to
`-count` to allow accessing the very first item.

For example (content_count = 5):
|              |first item|  |  |  |last item|
|positive index|       0  | 1| 2| 3|    4    |
|negative index|      -5  |-4|-3|-2|   -1    |

If negative indices are limited to be >= -4 the first item cannot be accessed
using negative indices.

Also, range limit of `pack_at` is removed for usability.

Reviewed-by: Marcel Hollerbach <mail@marcel-hollerbach.de>
Reviewed-by: YeongJong Lee <yj34.lee@samsung.com>
Differential Revision: https://phab.enlightenment.org/D8433