eo: add events to track the ownership status of an Eo object
Summary:
Some user code may want to track an object ownership in regard to whether it is
kept by just one owner or shared between many owners.

This is specially true for code provided by bindings to other programming
languages, where different kinds of resource management may take place.

The event `ownership,unique` is triggered whenever the object refcount goes
from two to one, as a signal that it has just one owner from now on.

The event `ownership,shared` is triggered whenever the object refcount goes
from one to two, as a signal that it has multiple owners from now on.
It will not trigger when further increasing the refcount to any value beyond
two.

We also add benchmarks for sharing (i.e. increasing the refcount) and them
unsharing objects, in order to evaluate the performance impact of this patch.

Reviewers: lauromoura, felipealmeida, cedric, bu5hm4n, segfaultxavi

Subscribers: #reviewers, woohyun, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D8678