mono-docs: Use full name in list of derived classes
In the EFL API we have several objects with the same name in different namespaces.
Thus it is important to use fully-qualified names throughout the docs, to avoid
confusion.
However, the default DocFX templates prefer using only simple class names.
The default templates have already been modified to use full names everywhere
except in the list of derived classes in the hierarchy section, where a DocFX
bug was hit.
This commit is a workaround to that bug (see
https://github.com/dotnet/docfx/issues/3769#issuecomment-485616064 )
so full names (with links) are used in the whole hierarchy.