docs: Fix genlist widget preview generation
Summary:
This was making "ninja doc" segfault for me.
If these function tables are not initialized they contain garbage, which leads
to the worst kind of undefined behavior.

Test Plan: Apply D8561 and try `ninja doc`. It was dumping core for me, and it does not do it anymore after this patch.

Reviewers: bu5hm4n, zmike, cedric

Reviewed By: bu5hm4n

Subscribers: #reviewers, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D8697